Add function to fix $PWD after a tmux restore
When starting a new window/pane with an explicit start-directory, tmux
fully resolves that directory. This breaks the tilde path shortening in
my prompt if `$HOME` contains symlinks, as zsh searches for `$HOME`, not
`$HOME:A`.

If `$HOME` contains symlinks and we're inside tmux, automatically run
this function at shell startup, so I don't have to do it myself.